[Android] Move CachedParams into ChromeFeatureList
Move instances of all subclasses of CachedFieldTrialParameter into
ChromeFeatureList.

CachedFlags are already there, along with MutableFlagsWithSafeParams
and MutableParamsWithSafeDefault.

This reduces the number of class initializers run when caching params.
It also lets us test that all CachedParams declared have been added to
the list of CachedParams (not included in this CL). And it lets us
test that the fieldtrial_testing_config.json matches
|defaultValueInTests| when this is added.

Cq-Include-Trybots: luci.chrome.try:android-internal-dbg
Bug: 373902543
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6055554
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Owners-Override: Andrew Grieve <agrieve@chromium.org>
Auto-Submit: Henrique Nakashima <hnakashima@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1389937}

+++ chrome/browser/auxiliary_search/junit/src/org/chromium/chrome/browser/auxiliary_search/AuxiliarySearchDonorUnitTest.java
@@ -64,7 +64,7 @@ public class AuxiliarySearchDonorUnitTest {
     public void testDefaultTtlIsNotZero() {
         assertNotEquals(0L, mAuxiliarySearchDonor.getDocumentTtlMs());
         assertEquals(
-                AuxiliarySearchUtils.DEFAULT_TTL_HOURS * 60 * 60 * 1000,
+                ((long) AuxiliarySearchUtils.DEFAULT_TTL_HOURS) * 60 * 60 * 1000,
                 mAuxiliarySearchDonor.getDocumentTtlMs());
     }
 
@@ -74,7 +74,7 @@ public class AuxiliarySearchDonorUnitTest {
     public void testConfiguredTtlCannotBeZero() {
         assertNotEquals(0L, mAuxiliarySearchDonor.getDocumentTtlMs());
         assertEquals(
-                AuxiliarySearchUtils.DEFAULT_TTL_HOURS * 60 * 60 * 1000,
+                ((long) AuxiliarySearchUtils.DEFAULT_TTL_HOURS) * 60 * 60 * 1000,
                 mAuxiliarySearchDonor.getDocumentTtlMs());
     }
